refactor: extract ClearSkyApp widget
